requirements.txt
streamlit
pypdf2
python-docx
openpyxlz
pandas
langchain-text-splitters
google-generativeai
faiss-cpu
python-dotenv
python-pptx
dotenv